+++ test/specs/scheme-payments-by-year.test.js
@@ -1,6 +1,6 @@
 import { test, expect } from '@playwright/test'
 import { expectPhaseBanner } from '../../utils/phase-banner-expect.js'
-import { expectNewTab } from '../../utils/new-page-link-expect.js'
+import { expectNewTab } from '../../utils/new-tab-link-expect.js'
 import { AccessibilityTest } from '../accessibility.test.js'
 import { securityTest } from '../security.test.js'
 import { expectRelatedContent } from '../../utils/related-content-expect.js'
