+++ utils/new-tab-link-expect.js
